Split IconLinkButton
