+++ docs/src/modules/common/AppBar.tsx
@@ -2,7 +2,8 @@ import * as React from 'react';
 import { BaseUIIcon } from 'docs-base/src/icons/BaseUI';
 import { GitHubIcon } from 'docs-base/src/icons/GitHub';
 import { SettingsIcon } from 'docs-base/src/icons/Settings';
-import { IconButton, IconLinkButton } from 'docs-base/src/design-system/IconButton';
+import { IconButton } from 'docs-base/src/design-system/IconButton';
+import { IconLinkButton } from 'docs-base/src/design-system/IconLinkButton';
 import { VersionSelector } from './VersionSelector';
 import packageJson from '../../../../package.json';
 import classes from './AppBar.module.css';
